+++ tests/testthat/test-api-repeats.R
@@ -57,7 +57,7 @@ for (cov in discrete_moderators) {
 
 qoi.cfg <- QoI_cfg$new(
     mcate = MCATE_cfg$new(cfgs = qoi.list),
-    vimp = VIMP_cfg$new(model_cfg = regression.cfg),
+    vimp = VIMP_cfg$new(),
     diag = Diagnostics_cfg$new(
         outcome = c("SL_risk", "SL_coefs", "MSE")
     )
@@ -66,6 +66,7 @@ qoi.cfg <- QoI_cfg$new(
 cfg <- HTE_cfg$new(
     treatment = trt.cfg,
     outcome = regression.cfg,
+    effect = regression.cfg,
     qoi = qoi.cfg
 )