+++ tests/testthat/test-vimp.R
@@ -9,7 +9,7 @@ d <- dplyr::tibble(
 ) %>% dplyr::mutate(y = 2 * cov1 + cov3 + rnorm(n))
 
 test_that("splitting ensures even number of splits with VIMP", {
-    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new(model_cfg = SLEnsemble_cfg$new())))
+    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new()))
     expect_message(
         make_splits(d, uid, .num_splits = 5),
         "`num_splits` must be even if VIMP is requested as a QoI. Rounding up."
@@ -19,14 +19,18 @@ test_that("splitting ensures even number of splits with VIMP", {
 
 test_that("VIMP ensures even number of splits", {
     d <- make_splits(d, uid, .num_splits = 5)
-    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new(model_cfg = SLEnsemble_cfg$new())))
+    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new()))
     expect_error(
-        calculate_vimp(d, w, y, cov1, cov2, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp),
+        calculate_vimp(
+            d, w, y, cov1, cov2, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
+        ),
         "Number of splits must be even to calculate VIMP."
     )
     attr(d, "num_splits") <- 4
     expect_error(
-        calculate_vimp(d, w, y, cov1, cov2, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp),
+        calculate_vimp(
+            d, w, y, cov1, cov2, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
+        ),
         "Number of splits is inconsistent."
     )
 })
@@ -34,8 +38,11 @@ test_that("VIMP ensures even number of splits", {
 test_that("VIMP works with weights", {
     d <- make_splits(d, uid, .num_splits = 4)
     d$w <- rexp(n, 1 / 0.9) + 0.1
-    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new(model_cfg = SLEnsemble_cfg$new())))
-    expect_error(result <- calculate_vimp(d, w, y, cov1, cov2, cov3, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp), NA)
+    attr(d, "HTE_cfg") <- HTE_cfg$new(qoi = QoI_cfg$new(vimp = VIMP_cfg$new()))
+    expect_error(result <- calculate_vimp(
+        d, w, y, cov1, cov2, cov3, 
+        .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
+    ), NA)
 
     expect_gt(result$estimate[1], result$estimate[2])
     expect_gt(result$estimate[3], result$estimate[2])
@@ -49,9 +56,12 @@ test_that("VIMP works with weights without sample splitting", {
     d <- make_splits(d, uid, .num_splits = 4)
     d$w <- rexp(n, 1 / 0.9) + 0.1
     attr(d, "HTE_cfg") <- HTE_cfg$new(
-        qoi = QoI_cfg$new(vimp = VIMP_cfg$new(sample_splitting = FALSE, model_cfg = SLEnsemble_cfg$new()))
+        qoi = QoI_cfg$new(vimp = VIMP_cfg$new(sample_splitting = FALSE))
     )
-    expect_error(result <- calculate_vimp(d, w, y, cov1, cov2, cov3, .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp), NA)
+    expect_error(result <- calculate_vimp(
+        d, w, y, cov1, cov2, cov3,
+        .VIMP_cfg = attr(d, "HTE_cfg")$qoi$vimp, .Model_cfg = SLEnsemble_cfg$new()
+    ), NA)
 
     expect_gt(result$estimate[1], result$estimate[2])
     expect_gt(result$estimate[3], result$estimate[2])